feat(depsync): add centralized dependency sync and integrate it into bump commands
