+++ internal/commands/bump/bump_plugins_test.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/indaco/sley/internal/commands/depsync"
 	"github.com/indaco/sley/internal/config"
 	"github.com/indaco/sley/internal/plugins"
 	"github.com/indaco/sley/internal/plugins/auditlog"
@@ -214,13 +215,13 @@ func TestSyncDependencies(t *testing.T) {
 	t.Run("nil checker returns nil", func(t *testing.T) {
 		registry := plugins.NewPluginRegistry()
 		dependencycheck.GetDependencyCheckerFn = func() dependencycheck.DependencyChecker { return nil }
-		err := syncDependencies(registry, version)
+		err := depsync.SyncDependencies(registry, version)
 		if err != nil {
 			t.Errorf("expected nil error, got %v", err)
 		}
 	})
 
-	// Note: syncDependencies uses type assertion to *DependencyCheckerPlugin
+	// Note: shared.SyncDependencies uses type assertion to *DependencyCheckerPlugin
 	// so mock implementations will be treated as disabled and return nil
 }
