+++ internal/commands/bump/common.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/indaco/sley/internal/clix"
+	"github.com/indaco/sley/internal/commands/depsync"
 	"github.com/indaco/sley/internal/config"
 	"github.com/indaco/sley/internal/operations"
 	"github.com/indaco/sley/internal/plugins"
@@ -121,7 +122,7 @@ func executePreBumpValidations(registry *plugins.PluginRegistry, newVersion, pre
 // used to avoid showing it twice in the dependency sync output.
 func executePostBumpActions(registry *plugins.PluginRegistry, newVersion, previousVersion semver.SemVersion, bumpType, bumpedPath string) error {
 	// Sync dependency files after updating .version
-	if err := syncDependencies(registry, newVersion, bumpedPath); err != nil {
+	if err := depsync.SyncDependencies(registry, newVersion, bumpedPath); err != nil {
 		return err
 	}
